Final Merge and lint errors
